feat: Add conditional execution of InfoSender in app.py
This commit adds functionality to app.py to conditionally execute the InfoSender script based on the presence of the --master-ip argument.

When running 'python -m app --master-ip http://localhost:18080', the InfoSender script is executed in a separate thread, registering and updating the worker information.

If no --master-ip argument is provided, the Flask app is started as usual.